+++ rusk/tests/services/mod.rs
@@ -10,6 +10,7 @@ pub mod conversion;
 //pub mod deploy;
 pub mod finalization;
 pub mod gas_behavior;
+pub mod http_propagate;
 pub mod init;
 pub mod mainnet;
 pub mod moonlight_stake;