+++ vm/tests/vm.rs
@@ -452,6 +452,9 @@ impl Circuit for PlonkTestCircuit {
 fn plonk_proof() {
     let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
+    let _plonk_version_guard = dusk_vm::host_queries::set_plonk_version(
+        dusk_core::plonk::PlonkVersion::current(),
+    );
 
     let pp = include_bytes!("./pp_test.bin");
     let pp = unsafe { PublicParameters::from_slice_unchecked(&pp[..]) };