+++ rusk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2026-02-27
+
 ### Changed
 
 - Update to edition 2024
@@ -31,6 +33,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add feature activation flags for `sha256` and `verify_kzg_proof` host queries [#3993]
 - Add feature activation flag for `secp256k1_recover` host query [#4012]
 - Fold phoenix refund check into `HARDFORK_AEGIS` feature flag
+
+## [1.5.0] - 2026-02-16
+
+### Changed
+
+- Set `FEATURE_PLONK_V2` activation heights for mainnet and testnet.
 
 ## [1.4.4] - 2026-01-30
 
@@ -502,7 +510,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#292]: https://github.com/dusk-network/rusk/issues/292
 [#290]: https://github.com/dusk-network/rusk/issues/290
 
-[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.4.4...HEAD
+[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.6.0...HEAD
+[1.6.0]: https://github.com/dusk-network/rusk-private/compare/0c61088614bcb515ed3abd955b738f4d375ce19c...dusk-rusk-1.6.0
+[1.5.0]: https://github.com/dusk-network/rusk-private/compare/dusk-rusk-1.4.4...0c61088614bcb515ed3abd955b738f4d375ce19c
 [1.4.4]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.4.3...dusk-rusk-1.4.4
 [1.4.3]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.4.2...dusk-rusk-1.4.3
 [1.4.2]: https://github.com/dusk-network/rusk/compare/dusk-rusk-1.4.1...dusk-rusk-1.4.2